requirements.txt
# BOOK-QI - Document Intelligence Platform
# ================================
# Django REST Framework Backend Requirements

# Core
djangorestframework>=3.14
django-cors-headers>=4.3

# Database - MySQL (set USE_SQLITE=true to fallback to SQLite)
mysqlclient>=2.0  # Optional: pip install mysqlclient (may need Visual Studio on Windows)

# Vector Store - ChromaDB
chromadb>=0.4

# AI & ML - Sentence Transformers for embeddings
sentence-transformers>=2.0

# Web Scraping - Selenium + BeautifulSoup
selenium>=4.0
beautifulsoup4>=4.12
webdriver-manager>=4.0

# Utilities
python-dotenv>=1.0
requests>=2.31

# Optional: Background Processing
# celery>=5.0   # Uncomment for async tasks
# redis>=4.0    # Uncomment for Celery broker